requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
motor>=3.1.1
pydantic>=1.10.7
python-dotenv>=1.0.0


# CodeBreak Game: Installation and Setup Guide with MongoDB Atlas

## Overview

CodeBreak is a multiplayer cyberpunk-themed survival game with real-time interaction between players. This guide will help you set up and run the game using MongoDB Atlas as the backend database.

## System Requirements

- Python 3.7 or higher
- Internet connection for MongoDB Atlas
- Sufficient disk space for game assets

## Installation Steps

### 1. Install Required Packages

Open a terminal or command prompt and run:

```bash
pip install fastapi uvicorn motor python-jose[cryptography] passlib[bcrypt] python-multipart websockets pygame requests python-dotenv
```

### 2. MongoDB Atlas Setup

1. **Create a MongoDB Atlas Account**:
   - Go to [MongoDB Atlas](https://www.mongodb.com/cloud/atlas) and sign up or log in
   - Create a new cluster (the free tier is sufficient for development)

2. **Configure Network Access**:
   - In the Atlas dashboard, go to "Network Access"
   - Click "Add IP Address"
   - Choose "Allow Access from Anywhere" (for development purposes) or add your specific IP
   - Click "Confirm"

3. **Create a Database User**:
   - Go to "Database Access" in Atlas
   - Click "Add New Database User"
   - Create a username and password (save these credentials)
   - Select "Read and write to any database" for user privileges
   - Click "Add User"

4. **Get Your Connection String**:
   - Go to "Databases" in Atlas dashboard
   - Click "Connect" on your cluster
   - Choose "Connect your application"
   - Copy the connection string (it will look like: `mongodb+srv://username:<password>@cluster0.xxxxx.mongodb.net/?retryWrites=true&w=majority`)
   - Replace `<password>` with your database user's password

### 3. Configure Environment

Create a `.env` file in the root directory with the following variables:

```
MONGO_URL=mongodb+srv://username:password@cluster0.xxxxx.mongodb.net/codebreak_db?retryWrites=true&w=majority
SECRET_KEY=your-secret-key-here
```

Make these changes:
- Replace the entire MongoDB URL with your Atlas connection string
- Replace `your-secret-key-here` with a strong random string for security

## Running the Game

### 1. Start the Backend Server

```bash
uvicorn server:app --reload --host 0.0.0.0 --port 8000
```

This starts the FastAPI server that handles authentication, leaderboard, and multiplayer features. The server will connect to your MongoDB Atlas database instead of a local instance.

### 2. Launch the Login Page

In a separate terminal:

```bash
python login.py
```

This opens the login screen where you can register a new account or log in with existing credentials.

### 3. Register and Play

- On the login screen, enter a username and password
- Click "REGISTER" to create a new account
- After registering, click "LOGIN" with the same credentials
- The game will automatically start after successful login

## Multiplayer Setup

To play with multiple players, you have two options:

### Option 1: Multiple Computers on Same Network

1. Start the server on one computer
2. Make sure the server's IP address is accessible to other computers
3. Update the server URL in the client code on other computers to point to the server's IP
4. Run the login page on each computer and connect to the game

### Option 2: Multiple Instances on One Computer

1. Start the server
2. Open multiple terminals
3. Run `python login.py` in each terminal
4. Register/login with different usernames for each instance

## In-Game Controls

- **Arrow Keys**: Move player
- **Space**: Melee attack
- **F**: Ranged attack
- **E**: Use equipped tool
- **C**: Open/close crafting menu
- **1-3**: Select crafting options (when crafting menu is open)
- **T**: Open chat
- **ESC**: Pause game / Close menus

## Troubleshooting

### Connection Issues

- Verify your internet connection is stable
- Check that your MongoDB Atlas cluster is running
- Ensure your IP address is in the allowed list in Atlas Network Access
- Verify the connection string in your `.env` file is correct
- Check that the FastAPI server is running and the port isn't blocked

### Game Crashes

- Check terminal for error messages
- Verify all required packages are installed
- Ensure game assets (sprites, sounds) are in the correct directories

### Authentication Problems

- If registration fails, the username might already exist
- If login fails, verify credentials are correct
- For persistent issues, try restarting the server

### MongoDB Atlas Issues

- Check the "Cluster" page in Atlas to ensure your cluster is active
- Verify your database user has the right permissions
- Review the Network Access settings to ensure your IP is allowed

## Advanced: Monitoring Database

- You can view your game data directly in the MongoDB Atlas dashboard
- Go to "Browse Collections" to see player records, leaderboards, and game sessions
- Atlas provides monitoring tools to track database performance

## Support

For additional help or to report issues, please contact the development team or submit an issue to the project repository.


uvicorn backend.server:app --host 0.0.0.0 --port 8000 --reload
